backend/requirements.txt
Flask
pytesseract
Pillow
opencv-python
ultralytics
torch